Conflicts:
	src/main/java/com/gsr/myschool/back/client/web/application/valueList/popup/AddValueListPresenter.java
	src/main/java/com/gsr/myschool/back/client/web/application/valueList/popup/AddValueTypePresenter.java

+++ src/main/java/com/gsr/myschool/server/business/core/PieceJustif.java
@@ -6,15 +6,15 @@ import javax.persistence.*;
 public class PieceJustif implements java.io.Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     @Column(unique = true)
     private String nom;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer newId) {
+    public void setId(Long newId) {
         this.id = newId;
     }
